refactor(memory): split MemoryStore into FactStore, RecallStore, GraphStore and ColumnStore facets (#1959)
The 22-method monolith becomes four facets: FactStore (write, by-id
lookup, deletion, count, list), RecallStore: FactStore (query_filtered,
query_excluding), GraphStore (edges, bounded scans, edge_count) and
ColumnStore (query_columnar). MemoryStore stays as their sum with a
blanket impl, so every existing S: MemoryStore caller compiles
unchanged; an out-of-tree implementor replaces impl MemoryStore with
the four facet impls — the methods did not change, they moved. That is
the one break, and it is the point: velesdb-memory goes 0.13.0 ->
0.14.0 (0.x cadence: breaking rides a minor).

MemoryService's struct bound relaxes to S: FactStore and every method
carries the bound of the facet its body actually reaches, derived by
compiler fixpoint rather than by hand — remember transitively needs
GraphStore because it wires entities, recall needs RecallStore, and a
partial backend simply does not have the methods it cannot serve. The
in-workspace backends (NativeStore, WasmStore) split mechanically in
the same commit, so no release window exposes a half-adapted trait.

The test doubles get what #1959 asked for: RecordingStore drops its
three recall/columnar stubs and ScanCutStore its three query
delegations — a test drifting into an unimplemented facet is now a
compile error instead of a plausible empty answer. The two
memory-bridge doubles shrink to FactStore alone.

A backend's capability gap also stops being billed to the caller:
MemoryError::Unsupported(&'static str) with ErrorCategory::Unsupported
replaces the InvalidFilter refusal in the default list(), and the
node/python/wasm/MCP adapters map the new category explicitly under
the existing ErrorCategory::ALL coverage guards.

The where-clause weight pushed two over-budget files the wrong way, so
each splits along the seam the guard prescribes: service.rs sheds its
graph half into service_graph.rs (1969 -> 1723 lines) and
memory_bridge.rs sheds context compilation into
memory_bridge_compile.rs (1530 -> 929, off the budget list entirely) —
in both, the budget seam and the facet seam coincide.

Verified: 1233 memory tests + 676 wasm tests green, no-default-features
lib green, clippy pedantic clean, rustdoc link diff empty vs develop.
velesdb-node compiles; its lib tests cannot link here (napi needs a
Node host) — CI covers them.

Closes #1959.
