+++ crates/velesdb-memory/src/migration/tests/diagnosis.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::GraphStore;
 
 // ---------------------------------------------------------------------------
 // GATE 3 — the inventory
