+++ crates/velesdb-memory/src/migration/tests/preservation.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::GraphStore;
 
 // ---------------------------------------------------------------------------
 // GATE 4 — preservation
