+++ crates/velesdb-memory/src/service.rs
@@ -46,7 +46,10 @@ use crate::model::{
 use crate::mutation::MutationObserver;
 #[cfg(feature = "persistence")]
 use crate::storage::NativeStore;
-use crate::storage::{is_reserved_key, strip_reserved_keys, MemoryStore, AUTO_DATE_FIELD};
+use crate::storage::{
+    is_reserved_key, strip_reserved_keys, ColumnStore, FactStore, GraphStore, RecallStore,
+    AUTO_DATE_FIELD,
+};
 
 /// [`MemoryService::recall_fused`] and its helpers — split out to keep this
 /// file under the crate's 500-NLOC-per-file budget, same pattern as
@@ -54,6 +57,11 @@ use crate::storage::{is_reserved_key, strip_reserved_keys, MemoryStore, AUTO_DAT
 /// shares full access to `MemoryService`'s private fields and methods.
 #[path = "fused_recall.rs"]
 mod fused_recall;
+
+/// The graph facet — `relate`/`forget`/hubs/`why`/`traverse` — same split,
+/// same reason; see `service_graph.rs`'s module doc.
+#[path = "service_graph.rs"]
+mod graph;
 
 #[cfg(feature = "persistence")]
 #[allow(dead_code)]
@@ -116,17 +124,20 @@ const ABOUT_RELATION: &str = "about";
 /// Local-first agent memory backed by a single `VelesDB` instance.
 ///
 /// Generic over the [`Embedder`] so production can use an on-device model while
-/// tests use a deterministic, network-free one, and over the [`MemoryStore`]
+/// tests use a deterministic, network-free one, and over the [`FactStore`]
 /// backend `S` so the same orchestration runs over the native, file-backed
 /// engine (the default — nothing changes for existing callers) or any other
-/// backend that implements the trait (e.g. an in-memory one for WASM).
+/// backend that implements the storage facets it uses (e.g. an in-memory one
+/// for WASM). Methods needing recall, graph, or columnar capability carry
+/// that facet as an extra bound — a partial backend simply does not have
+/// those methods (#1959).
 ///
 /// Two definitions, `persistence`-gated: the default type parameter itself
 /// references [`NativeStore`], which doesn't exist as a type at all without
 /// the feature, so a `persistence`-free build (e.g. `velesdb-wasm`) drops the
-/// default and every caller names its own [`MemoryStore`] backend explicitly.
+/// default and every caller names its own storage backend explicitly.
 #[cfg(feature = "persistence")]
-pub struct MemoryService<E: Embedder, S: MemoryStore = NativeStore> {
+pub struct MemoryService<E: Embedder, S: FactStore = NativeStore> {
     store: S,
     embedder: E,
     autograph: Option<crate::extract::DynExtractor>,
@@ -134,7 +145,7 @@ pub struct MemoryService<E: Embedder, S: MemoryStore = NativeStore> {
     generation_gate: parking_lot::RwLock<()>,
 }
 #[cfg(not(feature = "persistence"))]
-pub struct MemoryService<E: Embedder, S: MemoryStore> {
+pub struct MemoryService<E: Embedder, S: FactStore> {
     store: S,
     embedder: E,
     autograph: Option<crate::extract::DynExtractor>,
@@ -245,7 +256,7 @@ impl<E: Embedder> MemoryService<E, NativeStore> {
     }
 }
 
-impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
+impl<E: Embedder, S: FactStore> MemoryService<E, S> {
     /// Build a service directly over a `store` backend, bypassing
     /// [`Self::open`]'s filesystem-specific setup — the constructor a
     /// non-native backend (e.g. `velesdb-wasm`'s in-memory store) uses.
@@ -338,7 +349,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         fact: &str,
         links: &[Link],
         metadata: Option<&Metadata>,
-    ) -> Result<u64, MemoryError> {
+    ) -> Result<u64, MemoryError>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         self.remember_inner(fact, links, metadata, None, true)
     }
@@ -381,7 +395,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         links: &[Link],
         metadata: Option<&Metadata>,
         ttl_seconds: Option<u64>,
-    ) -> Result<u64, MemoryError> {
+    ) -> Result<u64, MemoryError>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         self.remember_inner(fact, links, metadata, ttl_seconds, true)
     }
@@ -397,7 +414,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         metadata: Option<&Metadata>,
         ttl_seconds: Option<u64>,
         run_autograph: bool,
-    ) -> Result<u64, MemoryError> {
+    ) -> Result<u64, MemoryError>
+    where
+        S: GraphStore,
+    {
         let fact = fact.trim();
         self.validate_write(fact, links, metadata, ttl_seconds)?;
         let fact_id = id::stable_id(fact);
@@ -467,7 +487,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         fact_id: u64,
         links: &[Link],
         existed_before: bool,
-    ) -> Result<(), MemoryError> {
+    ) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         let Err(cause) = self.relate_links(fact_id, links) else {
             return Ok(());
         };
@@ -493,7 +516,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// re-remembering, stalling every write behind a slow model is not. A
     /// disconnected queue (worker gone) falls back inline, so the graph
     /// keeps building even if the worker died.
-    fn autograph_if(&self, run: bool, fact_id: u64, fact: &str) {
+    fn autograph_if(&self, run: bool, fact_id: u64, fact: &str)
+    where
+        S: GraphStore,
+    {
         if !run {
             return;
         }
@@ -552,7 +578,7 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     }
 
     /// The total number of live tracked facts, internal entity hubs included
-    /// — the store's [`MemoryStore::count`], relayed for `memory_status`.
+    /// — the store's [`FactStore::count`], relayed for `memory_status`.
     #[must_use]
     pub fn fact_count(&self) -> usize {
         let _generation = self.enter_generation();
@@ -560,11 +586,14 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     }
 
     /// The total number of graph edges, when the backend can say —
-    /// [`MemoryStore::edge_count`], relayed for `memory_status`. `None`
+    /// [`GraphStore::edge_count`], relayed for `memory_status`. `None`
     /// means "cannot say", never "zero": the two answers tell a caller
     /// different things about `why()`.
     #[must_use]
-    pub fn edge_count(&self) -> Option<usize> {
+    pub fn edge_count(&self) -> Option<usize>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         self.store.edge_count()
     }
@@ -574,7 +603,7 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// resemblance to a query, and what resembles nothing you thought to
     /// ask stays invisible.
     ///
-    /// The store hands back raw pages ([`MemoryStore::list`]); the
+    /// The store hands back raw pages ([`FactStore::list`]); the
     /// visibility policy is applied here, once, for every backend: internal
     /// entity hubs are skipped unless `include_internal` (they are the
     /// graph's scaffolding, not the user's facts), reserved `_veles_*` keys
@@ -609,14 +638,17 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
 impl<E, S> MemoryService<E, S>
 where
     E: Embedder + Send + Sync + 'static,
-    S: MemoryStore + Send + Sync + 'static,
+    S: FactStore + Send + Sync + 'static,
 {
     /// The autograph worker's whole life, run on the spawned thread: ends
     /// when every sender is gone — i.e. when the handle's drop takes the
     /// sender back out of the service. Once the closing latch is up,
     /// still-queued jobs are SKIPPED: the exit pays for the job in flight,
     /// never for the queue.
-    fn autograph_worker_loop(&self, rx: &std::sync::mpsc::Receiver<AutographJob>) {
+    fn autograph_worker_loop(&self, rx: &std::sync::mpsc::Receiver<AutographJob>)
+    where
+        S: GraphStore,
+    {
         let mut skipped_on_close: u64 = 0;
         for job in rx {
             if self
@@ -673,7 +705,10 @@ where
     pub fn spawn_autograph_worker(
         self: &std::sync::Arc<Self>,
         capacity: usize,
-    ) -> Result<AutographWorkerHandle, MemoryError> {
+    ) -> Result<AutographWorkerHandle, MemoryError>
+    where
+        S: GraphStore,
+    {
         let (tx, rx) = std::sync::mpsc::sync_channel::<AutographJob>(capacity);
         {
             let mut guard = self.autograph_queue.tx.lock();
@@ -716,7 +751,7 @@ where
     }
 }
 
-impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
+impl<E: Embedder, S: FactStore> MemoryService<E, S> {
     /// Autograph one just-stored fact: read the entities, entity→entity edges
     /// and attributes it states, and wire them around it.
     ///
@@ -732,7 +767,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// The trade-off is that a persistently broken extractor degrades silently
     /// to plain `remember`. That is the right way round — losing structure is
     /// recoverable by re-remembering, losing the fact is not.
-    fn autograph(&self, fact_id: u64, fact: &str) {
+    fn autograph(&self, fact_id: u64, fact: &str)
+    where
+        S: GraphStore,
+    {
         let Some(extractor) = self.autograph.as_ref() else {
             return;
         };
@@ -789,7 +827,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// [`Self::remember_with_ttl`]'s pre-write pass (its only caller) —
     /// no re-check here, so the validation rule lives in exactly one
     /// place on this path.
-    fn relate_links(&self, fact_id: u64, links: &[Link]) -> Result<(), MemoryError> {
+    fn relate_links(&self, fact_id: u64, links: &[Link]) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         for link in links {
             self.store.relate(fact_id, link.target, &link.relation)?;
         }
@@ -827,7 +868,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         text: &str,
         extractor: &X,
         metadata: Option<&Metadata>,
-    ) -> Result<RememberedExtraction, MemoryError> {
+    ) -> Result<RememberedExtraction, MemoryError>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         let extraction = Self::extract_passage(text, extractor)?;
         self.store_extraction_inner(&extraction, metadata)
@@ -859,7 +903,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         &self,
         extraction: &crate::extract::Extraction,
         metadata: Option<&Metadata>,
-    ) -> Result<RememberedExtraction, MemoryError> {
+    ) -> Result<RememberedExtraction, MemoryError>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         self.store_extraction_inner(extraction, metadata)
     }
@@ -868,7 +915,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         &self,
         extraction: &crate::extract::Extraction,
         metadata: Option<&Metadata>,
-    ) -> Result<RememberedExtraction, MemoryError> {
+    ) -> Result<RememberedExtraction, MemoryError>
+    where
+        S: GraphStore,
+    {
         let mut entity_ids: HashMap<String, u64> = HashMap::new();
         let mut edges: HashSet<(u64, u64, String)> = HashSet::new();
         let outcome =
@@ -896,7 +946,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     ///
     /// # Errors
     /// Returns [`MemoryError`] if the store lookup fails.
-    pub fn entity_profile(&self, name: &str) -> Result<Option<EntityProfile>, MemoryError> {
+    pub fn entity_profile(&self, name: &str) -> Result<Option<EntityProfile>, MemoryError>
+    where
+        S: GraphStore,
+    {
         let _generation = self.enter_generation();
         let key = canonical_entity_name(name);
         if key.is_empty() {
@@ -928,10 +981,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// Scaffolding edges (`mentions`, and `about` for symmetry — a hub never
     /// has an outgoing `about`) are dropped: they point at the facts that
     /// tagged this entity, not at a statement *about* it.
-    fn outgoing_entity_relations(
-        &self,
-        id: u64,
-    ) -> Result<(Vec<EntityRelation>, bool), MemoryError> {
+    fn outgoing_entity_relations(&self, id: u64) -> Result<(Vec<EntityRelation>, bool), MemoryError>
+    where
+        S: GraphStore,
+    {
         let scanned = self
             .store
             .relations_bounded(id, crate::limits::MAX_ENTITY_SCAN_EDGES)?;
@@ -950,10 +1003,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
     /// [`Self::outgoing_entity_relations`]'s: `about` edges are dropped (they
     /// are the fact → hub half of the `about`/`mentions` pair), and `mentions`
     /// with them for symmetry.
-    fn incoming_entity_relations(
-        &self,
-        id: u64,
-    ) -> Result<(Vec<EntityRelation>, bool), MemoryError> {
+    fn incoming_entity_relations(&self, id: u64) -> Result<(Vec<EntityRelation>, bool), MemoryError>
+    where
+        S: GraphStore,
+    {
         let scanned = self
             .store
             .incoming_relations_bounded(id, crate::limits::MAX_ENTITY_SCAN_EDGES)?;
@@ -1018,7 +1071,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         relations: &[ExtractedRelation],
         entity_ids: &mut HashMap<String, u64>,
         edges: &mut HashSet<(u64, u64, String)>,
-    ) -> Result<(), MemoryError> {
+    ) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         for relation in relations {
             if validate_relation(&relation.predicate).is_err() {
                 continue;
@@ -1086,7 +1142,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         metadata: Option<&Metadata>,
         entity_ids: &mut HashMap<String, u64>,
         edges: &mut HashSet<(u64, u64, String)>,
-    ) -> Result<RememberedExtraction, MemoryError> {
+    ) -> Result<RememberedExtraction, MemoryError>
+    where
+        S: GraphStore,
+    {
         let mut ids = Vec::with_capacity(facts.len());
         let mut skipped_over_cap = 0;
         for fact in facts {
@@ -1127,7 +1186,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         entities: &[String],
         entity_ids: &mut HashMap<String, u64>,
         edges: &mut HashSet<(u64, u64, String)>,
-    ) -> Result<(), MemoryError> {
+    ) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         for entity in entities {
             // Skip blank or punctuation-only topics: they would persist as junk
             // hubs (`Entity: -`) yet can never carry a meaningful multi-hop link.
@@ -1146,7 +1208,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         entity: &str,
         entity_ids: &mut HashMap<String, u64>,
         edges: &mut HashSet<(u64, u64, String)>,
-    ) -> Result<(), MemoryError> {
+    ) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         let entity_id = self.entity_hub(entity, entity_ids)?;
         if entity_id == fact_id {
             return Ok(());
@@ -1170,7 +1235,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         to: u64,
         label: &str,
         edges: &mut HashSet<(u64, u64, String)>,
-    ) -> Result<(), MemoryError> {
+    ) -> Result<(), MemoryError>
+    where
+        S: GraphStore,
+    {
         if edges.insert((from, to, label.to_string())) {
             self.relate_inner(from, to, label)?;
         }
@@ -1288,7 +1356,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         query: &str,
         k: usize,
         filter: Option<&Metadata>,
-    ) -> Result<Vec<Recollection>, MemoryError> {
+    ) -> Result<Vec<Recollection>, MemoryError>
+    where
+        S: RecallStore,
+    {
         let _generation = self.enter_generation();
         self.recall_inner(query, k, filter)
     }
@@ -1298,7 +1369,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         query: &str,
         k: usize,
         filter: Option<&Metadata>,
-    ) -> Result<Vec<Recollection>, MemoryError> {
+    ) -> Result<Vec<Recollection>, MemoryError>
+    where
+        S: RecallStore,
+    {
         let query = query.trim();
         if query.is_empty() {
             return Ok(Vec::new());
@@ -1334,7 +1408,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         embedding: &[f32],
         k: usize,
         filter: Option<&Metadata>,
-    ) -> Result<Vec<(u64, f32, String)>, MemoryError> {
+    ) -> Result<Vec<(u64, f32, String)>, MemoryError>
+    where
+        S: RecallStore,
+    {
         match filter {
             // An include filter already excludes hubs: a hub's payload
             // carries only reserved keys (`content`, `_veles_hub`), and
@@ -1386,7 +1463,10 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         query: &str,
         k: usize,
         filters: &[ColumnFilter],
-    ) -> Result<Vec<Recollection>, MemoryError> {
+    ) -> Result<Vec<Recollection>, MemoryError>
+    where
+        S: ColumnStore + RecallStore,
+    {
         let _generation = self.enter_generation();
         let query = query.trim();
         if query.is_empty() || k == 0 {
@@ -1402,332 +1482,6 @@ impl<E: Embedder, S: MemoryStore> MemoryService<E, S> {
         }
         let embedding = self.embedder.embed(query)?;
         self.store.query_columnar(&embedding, k, filters)
-    }
-
-    /// Create a typed edge `from -> to`. Returns the edge id.
-    ///
-    /// Both endpoints are validated to exist first, so the tool reports an
-    /// unknown id as client input (`UnknownMemory`) rather than a generic
-    /// storage fault — and the graph never gains an edge dangling off a memory
-    /// that was never stored.
-    ///
-    /// A self-loop (`from == to`) is refused: it states nothing, and `why`
-    /// traverses it like any other edge, so it only adds noise to the
-    /// evidence trail. The same rule covers [`Self::remember`]'s `links`.
-    ///
-    /// # Errors
-    /// Returns [`MemoryError::InvalidRelation`] for a bad label,
-    /// [`MemoryError::SelfRelation`] if both endpoints are the same memory,
-    /// [`MemoryError::UnknownMemory`] if either endpoint is missing, or
-    /// a storage error if the edge cannot be created.
-    pub fn relate(&self, from: u64, to: u64, relation: &str) -> Result<u64, MemoryError> {
-        let _generation = self.enter_generation();
-        self.relate_inner(from, to, relation)
-    }
-
-    fn relate_inner(&self, from: u64, to: u64, relation: &str) -> Result<u64, MemoryError> {
-        validate_relation(relation)?;
-        if from == to {
-            return Err(MemoryError::SelfRelation(from));
-        }
-        self.ensure_exists(from)?;
-        self.ensure_exists(to)?;
-        self.store.relate(from, to, relation)
-    }
-
-    /// Remove the edge(s) `from -relation-> to`: [`Self::relate`]'s exact
-    /// undo (issue #1661), so a mistaken edge no longer costs the facts at
-    /// its endpoints. Neither the facts nor any entity hub are touched —
-    /// collecting an orphaned hub stays [`Self::forget`]'s job.
-    ///
-    /// Idempotent: an absent edge is `found: false`, not an error, so a
-    /// cleanup is replayable. It refuses exactly what `relate` refuses
-    /// (empty label, self-loop), and deliberately does NOT require the
-    /// endpoints to exist — the edge of a forgotten fact is already gone,
-    /// and reporting that as an error would break replay.
-    ///
-    /// Scope: the store does not distinguish an explicit edge from one the
-    /// autograph derived from a passage, so `unrelate` removes both alike.
-    /// To correct an autograph edge, prefer `forget` + `remember` of the
-    /// source fact — otherwise a later `remember` of the same passage can
-    /// rebuild the edge removed here.
-    ///
-    /// # Errors
-    /// Returns [`MemoryError::InvalidRelation`] for a bad label,
-    /// [`MemoryError::SelfRelation`] if both endpoints are the same memory,
-    /// or a storage error if lookup or removal fails.
-    pub fn unrelate(
-        &self,
-        from: u64,
-        to: u64,
-        relation: &str,
-    ) -> Result<UnrelateOutcome, MemoryError> {
-        let _generation = self.enter_generation();
-        validate_relation(relation)?;
-        if from == to {
-            return Err(MemoryError::SelfRelation(from));
-        }
-        let removed = self.remove_matching_edges(from, to, relation)?;
-        Ok(UnrelateOutcome {
-            found: removed > 0,
-            removed,
-        })
-    }
-
-    /// [`Self::unrelate`]'s removal pass: resolve `from`'s outgoing edges and
-    /// delete every one matching `(to, relation)` by its id, counting them.
-    fn remove_matching_edges(
-        &self,
-        from: u64,
-        to: u64,
-        relation: &str,
-    ) -> Result<usize, MemoryError> {
-        let mut removed = 0usize;
-        for edge in self.store.relations(from)? {
-            if edge.to == to
-                && edge.relation == relation
-                && self.store.unrelate_from(from, edge.id)?
-            {
-                removed += 1;
-            }
-        }
-        Ok(removed)
-    }
-
-    /// Forget (delete) the memory with `fact_id`. Returns whether a memory
-    /// actually existed under that id — the underlying store's `delete` is a
-    /// silent no-op on an unknown id (matching most backends' idempotent
-    /// delete semantics), which is indistinguishable from a real deletion
-    /// unless existence is checked first. Every surface that exposes
-    /// `forget` (MCP, Node, WASM, Python) forwards this so a caller can tell
-    /// "I removed something" from "that id was a typo".
-    ///
-    /// The delete always runs, even when `get` reports the id absent: `get`
-    /// filters TTL-expired facts, and an expired-but-unpurged row must still
-    /// be reclaimed (the caller is told `false` — the memory was already
-    /// gone from its perspective). Existence check and delete are two store
-    /// calls, not one atomic operation: two concurrent forgets of one id may
-    /// both report `true`.
-    ///
-    /// # Errors
-    /// Returns [`MemoryError`] if the existence check or the deletion fails.
-    pub fn forget(&self, fact_id: u64) -> Result<bool, MemoryError> {
-        let _generation = self.enter_generation();
-        let found = self.store.get(fact_id)?.is_some();
-        // Read the fact's hubs BEFORE the delete: afterwards its edges are gone
-        // and there is no way back to the entities it created.
-        let hubs = self.hubs_linked_from(fact_id)?;
-        self.store.delete(fact_id)?;
-        self.collect_orphan_hubs(&hubs)?;
-        Ok(found)
-    }
-
-    /// The entity hubs `fact_id` points at.
-    ///
-    /// Hubs are recognised by the reserved [`HUB_FIELD`] marker rather than by
-    /// the edge label, so a caller's own `relate` to a hub is seen too.
-    fn hubs_linked_from(&self, fact_id: u64) -> Result<Vec<u64>, MemoryError> {
-        let mut hubs = Vec::new();
-        for edge in self.store.relations(fact_id)? {
-            if self.is_hub(edge.to)? {
-                hubs.push(edge.to);
-            }
-        }
-        Ok(hubs)
-    }
-
-    /// Delete every hub in `hubs` that no surviving fact mentions any more.
-    ///
-    /// An entity outlives the fact that introduced it as long as another fact
-    /// still refers to it — forgetting "Theo is 15" must not erase Theo while
-    /// "Theo has a sister" is still stored. Only a hub whose every `mentions`
-    /// target is gone is itself removed, so entities do not accumulate as
-    /// unreachable scaffolding once the facts behind them are retracted.
-    fn collect_orphan_hubs(&self, hubs: &[u64]) -> Result<(), MemoryError> {
-        for &hub in hubs {
-            if !self.hub_still_mentioned(hub)? {
-                self.store.delete(hub)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Whether anything alive still needs `hub`.
-    ///
-    /// Two references count, and the second is why this reads BOTH
-    /// directions (issue #1662):
-    ///
-    /// - an outgoing `mentions` edge to a live fact — the pair
-    ///   [`Self::wire_entities`] writes, the ordinary case;
-    /// - an incoming edge from a live NON-HUB fact — what a caller's own
-    ///   `relate` writes, and it writes one direction only. Relating a fact
-    ///   to a hub is reachable (`entity()` hands out the hub id), so reading
-    ///   outgoing edges alone swept hubs from under live callers' edges,
-    ///   losing them in silence.
-    ///
-    /// Incoming edges from another HUB are deliberately ignored: hub↔hub
-    /// edges exist (`wire_relations` writes them), and counting them would
-    /// let two hubs keep each other alive forever — a leak whose outcome
-    /// depends on collection order, which is worse than the bug being fixed.
-    fn hub_still_mentioned(&self, hub: u64) -> Result<bool, MemoryError> {
-        for edge in self.store.relations(hub)? {
-            if edge.relation == MENTIONS_RELATION && self.store.get(edge.to)?.is_some() {
-                return Ok(true);
-            }
-        }
-        self.hub_has_live_referent(hub)
-    }
-
-    /// Whether a live non-hub fact points AT `hub` — see
-    /// [`Self::hub_still_mentioned`] for why hub→hub edges do not count.
-    fn hub_has_live_referent(&self, hub: u64) -> Result<bool, MemoryError> {
-        for edge in self.store.incoming_relations(hub)? {
-            if self.store.get(edge.from)?.is_some() && !self.is_hub(edge.from)? {
-                return Ok(true);
-            }
-        }
-        Ok(false)
-    }
-
-    /// Whether `id` is an entity hub (carries the reserved [`HUB_FIELD`]).
-    fn is_hub(&self, id: u64) -> Result<bool, MemoryError> {
-        Ok(self
-            .store
-            .get_metadata(id)?
-            .is_some_and(|meta| meta.contains_key(HUB_FIELD)))
-    }
-
-    /// Explain a `decision`: find the best-matching memory (optionally scoped to
-    /// a metadata `filter`, e.g. the current project), then walk its typed links
-    /// up to `max_hops` away — fusing the vector, `ColumnStore`, and graph facets.
-    ///
-    /// Returns an empty [`Explanation`] when nothing matches the decision.
-    ///
-    /// # Errors
-    /// Returns [`MemoryError`] if recall or graph traversal fails.
-    pub fn why(
-        &self,
-        decision: &str,
-        max_hops: usize,
-        filter: Option<&Metadata>,
-    ) -> Result<Explanation, MemoryError> {
-        let _generation = self.enter_generation();
-        let decision = decision.trim();
-        if decision.is_empty() {
-            return Ok(Explanation::default());
-        }
-        reject_reserved_keys(filter)?;
-        let embedding = self.embedder.embed(decision)?;
-        let seeds = self.search(&embedding, 1, filter)?;
-        let Some((seed_id, _score, seed_content)) = seeds.into_iter().next() else {
-            return Ok(Explanation::default());
-        };
-        self.traverse(seed_id, seed_content, max_hops)
-    }
-
-    /// Breadth-first walk over outgoing links from `seed_id`, collecting nodes
-    /// and edges up to `max_hops` away.
-    fn traverse(
-        &self,
-        seed_id: u64,
-        seed_content: String,
-        max_hops: usize,
-    ) -> Result<Explanation, MemoryError> {
-        let mut explanation = Explanation {
-            nodes: vec![MemoryNode {
-                id: seed_id,
-                content: seed_content,
-                hop: 0,
-            }],
-            edges: Vec::new(),
-            truncated: false,
-        };
-        let mut visited: HashSet<u64> = HashSet::from([seed_id]);
-        let mut frontier = vec![seed_id];
-        let mut next: Vec<u64> = Vec::new();
-        'hops: for hop in 1..=max_hops {
-            next.clear();
-            for node_id in frontier.drain(..) {
-                // Both width budgets, checked here AND inside `expand`: this
-                // check alone would let the expansion that crosses the line
-                // finish its node — up to MAX_WHY_NODE_DEGREE nodes past the
-                // "ceiling", which a review measured at 522 of a promised 500.
-                if explanation.nodes.len() >= crate::limits::MAX_WHY_NODES
-                    || explanation.edges.len() >= crate::limits::MAX_WHY_EDGES
-                {
-                    // Unexpanded frontier work remained — the response is a
-                    // partial view and must SAY so (#1820); whether the rest
-                    // held anything unseen is exactly what the budget forbids
-                    // finding out, so the cautious true is the honest one.
-                    explanation.truncated = true;
-                    break 'hops; // width budget spent — depth left in max_hops is moot
-                }
-                self.expand(node_id, hop, &mut explanation, &mut visited, &mut next)?;
-            }
-            if next.is_empty() {
-                break;
-            }
-            std::mem::swap(&mut frontier, &mut next);
-        }
-        Ok(explanation)
-    }
-
-    /// Expand a single node: enqueue unseen targets and record edges, following
-    /// at most [`crate::limits::MAX_WHY_NODE_DEGREE`] outgoing edges — an entity
-    /// hub's degree scales with the whole store, so an unbounded walk here would
-    /// dump its entire neighborhood into one response (issue #1743). An edge is
-    /// only recorded once its target is a resolved node, so the subgraph never
-    /// contains an edge pointing at a node absent from `nodes` (e.g. a forgotten
-    /// target whose edge outlived it).
-    fn expand(
-        &self,
-        node_id: u64,
-        hop: usize,
-        explanation: &mut Explanation,
-        visited: &mut HashSet<u64>,
-        next: &mut Vec<u64>,
-    ) -> Result<(), MemoryError> {
-        // The bounded read pushes the per-node budget into the store's own
-        // index scan: the old full fetch materialized a super-node's whole
-        // degree before `.take()` could apply — O(store size) transient
-        // allocation at a single hop, the cost half of #1743 that #1820
-        // closes. The store also reports whether the degree exceeded the
-        // budget, which is what makes the cut OBSERVABLE.
-        let bounded = self
-            .store
-            .relations_bounded(node_id, crate::limits::MAX_WHY_NODE_DEGREE)?;
-        if bounded.truncated {
-            explanation.truncated = true;
-        }
-        for edge in bounded.edges {
-            // The budgets are ceilings, not suggestions: once either is spent,
-            // this node's expansion stops MID-NODE rather than finishing. The
-            // caller's check between nodes cannot provide that — an expansion
-            // that crosses the line would otherwise add its whole degree.
-            if explanation.nodes.len() >= crate::limits::MAX_WHY_NODES
-                || explanation.edges.len() >= crate::limits::MAX_WHY_EDGES
-            {
-                // An edge was in hand and not followed — an exact cut, not
-                // a conservative one.
-                explanation.truncated = true;
-                break;
-            }
-            let target = edge.to;
-            if !visited.contains(&target) {
-                let Some((content, _embedding)) = self.store.get(target)? else {
-                    continue; // target no longer exists → drop the dangling edge too
-                };
-                visited.insert(target);
-                explanation.nodes.push(MemoryNode {
-                    id: target,
-                    content,
-                    hop,
-                });
-                next.push(target);
-            }
-            explanation.edges.push(edge);
-        }
-        Ok(())
     }
 }
