+++ crates/velesdb-node/README.md
@@ -325,4 +325,4 @@ Questions: contact@wiscale.fr.
 
 ---
 
-`velesdb-node v0.13.0` (npm `@wiscale/velesdb-memory-node@0.13.0`) · Last updated: 2026-07-30 · Applies to: velesdb-core 5.1.0 · [Report a docs error](https://github.com/cyberlife-coder/VelesDB/issues)
+`velesdb-node v0.14.0` (npm `@wiscale/velesdb-memory-node@0.14.0`) · Last updated: 2026-08-19 · Applies to: velesdb-core 5.1.0 · [Report a docs error](https://github.com/cyberlife-coder/VelesDB/issues)
