+++ docs/guides/CONTEXT_COMPILER.md
@@ -841,4 +841,4 @@ skill teaches an agent the full workflow — including when *not* to compress.
 
 ---
 
-Last updated: 2026-07-25 · Applies to: velesdb-memory 0.13.0
+Last updated: 2026-07-25 · Applies to: velesdb-memory 0.14.0
